pyproject.toml
[project]
name = "python-spider"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "beautifulsoup4==4.13.4",
    "pycryptodome>=3.23.0",
    "pyexecjs>=1.5.1",
    "requests==2.31.0",
]
